requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
loguru==0.7.0
pydantic==1.10.9
pandas==2.0.2
sqlalchemy==2.0.16
psycopg2-binary==2.9.6